+++ src/widgets/containers/center_middle.rs
@@ -8,6 +8,8 @@ pub struct CenterMiddle {
 }
 
 impl CenterMiddle {
+    crate::delegate_ident_methods!(inner);
+
     pub fn new() -> Self {
         Self {
             inner: Container::new(),