+++ src/widgets/containers/middle.rs
@@ -8,6 +8,8 @@ pub struct Middle {
 }
 
 impl Middle {
+    crate::delegate_ident_methods!(inner);
+
     pub fn new() -> Self {
         Self {
             inner: Container::new(),