+++ src/widgets/containers/right.rs
@@ -8,6 +8,8 @@ pub struct Right {
 }
 
 impl Right {
+    crate::delegate_ident_methods!(inner);
+
     pub fn new() -> Self {
         Self {
             inner: Container::new(),